Changes to the project structure
