+++ src/helpers/request.helper.js
@@ -1,4 +1,4 @@
-'user strict';
+'use strict';
 
 const requestHelper = {
   getRequestId(req) {
