+++ src/helpers/response.helper.js
@@ -1,4 +1,4 @@
-'user strict';
+'use strict';
 
 //
 // internal module
