--- src/microservice/helpers/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const ErrorHelper = require('./modules/error.helper');
-const requestHelper = require('./modules/request.helper');
-const responseHelper = require('./modules/response.helper');
-
-module.exports = {
-  ErrorHelper,
-  requestHelper,
-  responseHelper
-};
