--- src/microservice/notes.microservice.js
@@ -1,26 +1,0 @@
-'use strict';
-
-//
-// exnternal modules
-const express = require('express');
-const bodyParser = require('body-parser');
-
-//
-// configure microservice with express
-const microservice = express();
-microservice.use(bodyParser.json());
-
-//
-// internal modules
-const notesController = require('./controllers/notes.controller');
-
-//
-// routes
-microservice.get('/notes/:id', notesController.getById);
-microservice.put('/notes/:id', notesController.update);
-microservice.delete('/notes/:id', notesController.delete);
-
-microservice.get('/notes', notesController.getAll);
-microservice.post('/notes', notesController.create);
-
-module.exports = microservice;
