+++ tests/002-get-all.test.js
@@ -8,7 +8,7 @@ const { expect } = require('chai');
 //
 // internal modules
 const helperTest = require('./helper.test');
-const api = require('../src/microservice/notes.microservice');
+const api = require('../src/microservice');
 
 describe('Get all notes', () => {
   const request = supertest(api);
