+++ tests/003-get-by-id.test.js
@@ -8,7 +8,7 @@ const { expect } = require('chai');
 //
 // internal modules
 const helperTest = require('./helper.test');
-const api = require('../src/microservice/notes.microservice');
+const api = require('../src/microservice');
 
 describe('Get note', () => {
   const request = supertest(api);
